refactor: isolate agent functionalities
The functionalities provided by the agent are implemented
directly in the main function. This limits reuse.
With this commit we isolate agent functionalities in a single
function called `run_agent` so that it can be leveraged in
multiple scenarios (like unit tests).
