--- libthymesisflow/include/agent.c
@@ -1,16 +1,0 @@
-/// @file server.h
-#if !defined(THYMESISFLOW_CMD_H)
-#define THYMESISFLOW_CMD_H
-
-#include <sys/socket.h>
-
-#ifndef MOCK
-#include "stats.h"
-#endif
-
-/*! \def BACKLOG
- * \brief max incoing connections
- */
-#define BACKLOG 5
-
-#endif
